checking out this new middleware that does a lot of the functions for us
