+++ server/plaid/plaidRouter.js
@@ -5,6 +5,7 @@ const qs = require('./plaidModel.js');
 const data = require('./data.js');
 
 const checkAccessToken = require("./getAccessToken-middleware.js");
+const webhookMiddle = require('./webhook-middleware.js');
 
 const router = express.Router();
 
@@ -74,7 +75,7 @@ router.post('/token_exchange', publicTokenExists, async (req, res) => {
 });
 
 //This is comming from PLAID, res.send or any variation will just be sending to plaid
-router.post('/webhook', async (req,res)=>{
+router.post('/webhook', webhookMiddle, async (req,res)=>{
   const body = req.body;
 
   
@@ -84,41 +85,41 @@ router.post('/webhook', async (req,res)=>{
 
     try{
 
-      //This is the item_id generated by Plaid. NOT the id column in the item table.NOT the primary key of that table
-      const item_id = body.item_id;
+      // //This is the item_id generated by Plaid. NOT the id column in the item table.NOT the primary key of that table
+      // const item_id = body.item_id;
   
-      //pgItemId holds the id that IS the primary key of the item table.
-      const pgItemId = await qs.WEB_get_pg_itemid(item_id)
+      // //pgItemId holds the id that IS the primary key of the item table.
+      // const pgItemId = await qs.WEB_get_pg_itemid(item_id)
 
-      //This is getting the userID for later. We need to get the userID from the item_id plaid gave us to get the EXACT user
-      const userID = await qs.WEB_get_userID(item_id)
+      // //This is getting the userID for later. We need to get the userID from the item_id plaid gave us to get the EXACT user
+      // const userID = await qs.WEB_get_userID(item_id)
   
       //This basically inserts a row that says that the transaction download into our db for this Item, essenially this user, has begun
-      const InsertionStart = await qs.WEB_track_insertion(pgItemId.id, 'inserting')
+      // const InsertionStart = await qs.WEB_track_insertion(pgItemId.id, 'inserting')
   
-      console.log('THE INSERTION BEGINNING', InsertionStart)
+      // console.log('THE INSERTION BEGINNING', InsertionStart)
   
-      //We just get the access token based on the ItemID plaid gave us to make sure we are accessing the transactions for that EXACT set of credentials
-      const {access_token} = await qs.WEB_get_accessToken(item_id)
+      // //We just get the access token based on the ItemID plaid gave us to make sure we are accessing the transactions for that EXACT set of credentials
+      // const {access_token} = await qs.WEB_get_accessToken(item_id)
 
-      //this is the current day
-      var end = (new Date()).toISOString().replace(/-/g, '-').split('T')[0]
-      //all three of these are to get 30 days behind the current day
-      var today = new Date()
-      var start = new Date().setDate(today.getDate()-30)
-      var start = (new Date(start)).toISOString().replace(/-/g, '-').split('T')[0]
+      // //this is the current day
+      // var end = (new Date()).toISOString().replace(/-/g, '-').split('T')[0]
+      // //all three of these are to get 30 days behind the current day
+      // var today = new Date()
+      // var start = new Date().setDate(today.getDate()-30)
+      // var start = (new Date(start)).toISOString().replace(/-/g, '-').split('T')[0]
       
-      console.log("DATE RANGE FOR TRANSACTIONS",start, end)
+      // console.log("DATE RANGE FOR TRANSACTIONS",start, end)
       
       //This is us getting the transactions 
-      const {transactions} = await client.getTransactions(access_token, start, end);
+      const {transactions} = await client.getTransactions(body.access_token, body.start, body.end);
 
       //This is a more refined version of what I had before on line 54. 
-     const done = await qs.WEB_insert_transactions(transactions, userID.id)
+     const done = await qs.WEB_insert_transactions(transactions, body.userID.id)
 
      //basically makes sure that the notification that the download is complete waits on it actually completing.
      if(done){ 
-       const InsertionEnd = await qs.WEB_track_insertion(pgItemId.id, 'done')
+       const InsertionEnd = await qs.WEB_track_insertion(body.pgItemId.id, 'done')
        console.log('THE INSERTION ENDING', InsertionEnd)
      }
